Overflow protection of the response buffer
When receiveing data and the remote sends more data than what we expect,
the response buffer could be overrun and corrupt the stack.

Send the maximum buffer length to lowLevelWriteRead(),
which declares a local buffer[PMAC_MAXBUF_] where data is received.
When the response is longer than expected, it is clipped and a warning
written.