sshDriver.cpp: debugStrPrintEscapedNL()
Add the function debugStrPrintEscapedNL() to print the escaped strings.
This function is inspired by epicsStrPrintEscaped().
The sshDriver is unaware of EPICS, so we do not want to use <epicsString.h>,
but if we want to save some bytes, we could use that function as well.

The main purpose is to get away with the epicsThreadSleep(1.0) and have a better
detection when gpascii is ready.
This will be done in another commit, for now we want to debug the communication.