Display the CPU temperature in the powerPMAC screen
When the cooling slits of a PowerBrick LV are covered, the CPU board gets
overheated and does not work reliable.
Add a PV, read it from the PMAC and show it on the powerPMAC screen,

There are actually 2 variables, one in ASCII and one having the values
as a float. The ASCII shows the raw output from the PMAC, it may be removed
later