Faster connection with gpascii: SSHDriver::syncInteractive
When gpascii is started, the original intention was to wait 1 second,
to allow gpascii to start up and be ready.

On the PMAC here the start is typically much faster, around 140 msec.
Introduce a sync method to detect when gpascii is running:
Send a newline and wait for the \006
Repeat it, if needed